backend/requirements.txt
# Flask及其扩展
Flask==2.2.5
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.4
python-socketio==5.8.0
python-dotenv==0.21.1

# 数据库
sqlalchemy==2.0.41
pymysql==1.1.0
cryptography==41.0.7
redis==5.0.1

# AI相关
openai==1.3.5
langchain==0.0.27
python-multipart==0.0.6
numpy==1.21.6
opencv-python==4.5.5.64
websockets==11.0.3

# 工具包
python-jose==3.3.0
passlib==1.7.4
requests==2.31.0
pytest==7.4.3
gunicorn==21.2.0